AddressTextField refactored
